requirements.txt
pywin32==306
psutil==5.9.5
pystray==0.19.4
email-validator==2.0.0
GPUtil==1.4.0
keyring==24.2.0
google-auth==2.22.0
google-api-python-client==2.98.0
google-auth-oauthlib==1.0.0
customtkinter==5.2.0
CTkListbox==0.9